apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.1.1
boto3==1.35.24
httpx==0.27.2
argon2-cffi==23.1.0
cryptography==43.0.1
orjson==3.10.7
pydantic-settings==2.5.2
